adding EmployeeForm to App.vue
firstly thank you very much for this tutorial tania - it was fun and educational.

secondly, please ignore this if leaving out these steps was an intentional learning opportunity for the learner. 

lastly, in case not here is the rough steps I needed to run through to actually get this stitched together. I haven't assumed to be able to write in your voice so this may not be in a mergeable voice.