made labels explicitly associated with their inputs elements for accessibility (#81)
in the tutorial there is a accessibility anti-pattern, that can be easily fixed. since this is an instructive tutorial on making webpages I think it is important for it to represent inclusive design, especially when it is a very small change.

In the article, the labels arent explicitly associated with their inputs. visually impaired and blind users might not be able to always tell which label is associated with which input, explicitly associating will make sure that their screen reader will present the input field together with its corresponding label. This also makes it easier to use for speech input users.